refactor: 시군구 검색 시그니처 변경
- 메소드 명 변경
- RestDocs 명세 변경
- DTO 패키지 내부로 이동
