+++ src/main/java/org/guzzing/studay_server/region/repository/RegionJpaRepository.java
@@ -3,9 +3,11 @@ package org.guzzing.studay_server.region.repository;
 import java.util.List;
 import org.guzzing.studay_server.region.model.Region;
 import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.data.jpa.repository.Query;
 
 public interface RegionJpaRepository extends JpaRepository<Region, Long>, RegionRepository {
 
-    List<String> findBySido(final String sido);
+    @Query("select distinct(r.sigungu) from Region r where r.sido = :sido")
+    List<String> findSigunguBySido(final String sido);
 
 }
