+++ src/main/java/org/guzzing/studay_server/region/repository/RegionRepository.java
@@ -5,7 +5,6 @@ import org.springframework.data.jpa.repository.Query;
 
 public interface RegionRepository {
 
-    @Query("select distinct(r.sigungu) from Region r where r.sido = :sido")
-    List<String> findBySido(final String sido);
+    List<String> findSigunguBySido(final String sido);
 
 }
