+++ src/main/java/org/guzzing/studay_server/region/service/RegionService.java
@@ -2,6 +2,7 @@ package org.guzzing.studay_server.region.service;
 
 import java.util.List;
 import org.guzzing.studay_server.region.repository.RegionRepository;
+import org.guzzing.studay_server.region.service.dto.SigunguResult;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
@@ -16,7 +17,7 @@ public class RegionService {
     }
 
     public SigunguResult findSigungusBySido(final String sido) {
-        List<String> sigungu = regionRepository.findBySido(sido);
+        List<String> sigungu = regionRepository.findSigunguBySido(sido);
         return SigunguResult.from(sido, sigungu);
     }
 }
